requirements.txt
coverage==4.5.1
Django==1.11.2
et-xmlfile==1.0.1
jdcal==1.4
pika==0.12.0
openpyxl==2.3.2
Pillow==5.2.0
psycopg2-binary==2.7.5
reportlab==3.5.5
django-analytical==2.4.0
django-localflavor==1.6
django-statici18n==1.8.2
factory_boy==2.8.1
django-dotenv==1.4.2
pendulum==1.2.0
djangorestframework==3.8.2
gitpython==2.1.9
requests==2.19.1
django-bootstrap3==10.0.1
django-ordered-model==2.1.0
django-waffle==0.14.0
django-redis==4.9.0
beautifulsoup4==4.6.3
pypom==2.1.0
selenium==3.11.0
python-magic==0.4.15
django-ajax-selects==1.7.0
celery==4.2.0
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-notifications-hq==1.5.0
django-filter==2.0.0
django-hijack==2.1.10
django-hijack-admin==2.1.10


-r ./osis_common/requirements.txt
-r ./internship/requirements.txt
-r ./partnership/requirements.txt
